.bash_logout: Prefer clear_console and tput clear
This commit switches from using the "clear" command to "clear_console"
and "tput clear". The clear_console command was designed specifically
for clearing the screens of consoles, so this is preferred. The tput
command is used as a fallback as it is defined by POSIX, and should be
available nearly everywhere.