vimrc: Use two spaces for Latex indentation
Eight-column tabs are too wide for most Latex documents. Latex is a
markup language, not a programming language. Because of this, highly
nested language structures are common and not undesirable. The use of
tabs causes indentation to take up most of the visual space of the
document instead of the body text of the document.